Brexit and Immigration: How Immigration affect the political climate in the UK
The British public voted to leave the European Union (EU) in the famous referendum in 2016. The result
was very impressive as the British public voted for leaving the EU with 51.9 percent against 48.1 percent
for remain. Previous researches show that the main factor that impacts the referendum’s results is immigration,
as the British voters want to stop the immigrants’ waves from the EU. The central question is
how immigration affects the political climate in the UK. In the theoretical background, I argue that the
social identity and culture participate in forming this vision about immigration among individuals, and then
affects their choices. For analysis, I investigated how immigration could affect the political parties’ selection
in 2015 (one year ahead of the referendum), by using the ESS dataset. I used multinomial logistic regression
with variables to measure immigration acceptance and how far this affects the political parties’ choice. The
results show that the individuals, who consider immigrants making the UK a worse place to live, have a high
tendency to vote for political parties that support Brexit and stop immigration such as the UK independence
party and the Conservative party.

To read the project, please check the link https://www.researchgate.net/publication/344943690_Brexit_and_Immigration_How_Immigration_affect_the_political_climate_in_the_UK